[WIP][packages/itk] Disable ITK modules and rename module
Disable some of the ITK modules explicitly with CMake flags as they
appear to cause some build issues.

Rename the former 'insight-toolkit' Guile module to 'itk', so it does
not cause confusion with the package name 'insight-toolkit'. Guix
convention is afterall that the package name reflects the upstream's
project name.