[ENH][packages/teem] Add Teem package definition
[ENH][packages/teem] Add build parameters for teem

[BUG][packages/teem] Fix missing TeemConfig.cmake

TeemConfig.cmake was not installed in lib/cmake/Teem/. Instead, it was
installed in lib/. Also, remove obsolete code.

[BUG][packages/teem] Fix failure to find TeemLibraryDepends.cmake

[COMP][packages/teem] Add Teem installation location

Dependencies are expecting Teem's CMake files in different locations.
Updating Teem's patch instead of applying them using phases appears to
the package author as a cleaner way to solve this problem.

[COMP][packages/teem] Add an additional Teem installation location

This commit fixes a compile error where CMake complains that
TeemUse.cmake was not found.

[COMP][packages/teem] Remove vtkaddon as dependency

[WIP][packages/teem] Change patch number to avoid conflict