[ENH][packages/itk] Make remote modules private
This commit replaces `(define-public)` by `(define)` for the remote
modules of `itk-slicer`. This effectively makes the remote modules as
private packages as they are not exposed. It does not make sense to
have these as public packages as they are not a standalone package,
but more like a part of `itk-slicer`